feat(encode): always return slice of Uint8Array, return next offset
This is much more useful, as it provides more information in the return
value than previously, with all of the old information still available.

The old return values gave you one effective piece of information which
was the length of bytes the VarInt took to encode. If you wanted the
bytes the VarInt is comprised of you needed to make your own slice with
the offset, and the bytes, into buf. If you wanted to know the next
offset you also had to do some math.

The new return values give you three effective pieces of information;
the VarInt encoded byte length (`[0].length`), the actual VarInt bytes
(`[0]`) and also the next offset ([1]). This makes for a much better
API. However it is a...

BREAKING CHANGE:

 - The returned `Uint8Array` from `encode()` is now a `slice()` of
   `buf`. Code should be updated to not expect the input `buf` to be the
   same as the output `Uint8Array`
 - The returned `number` is now the new offset, not the length of
   encoded bytes. To get the length of encoded bytes, instead check the
   length of the returned `Uint8Array`.