feat(decode): return the new offset instead of byte length
For better symmetry with `encode()`, `decode()` and `decode32()` now
return the new offset from which to continue decoding. This makes more
sense from not just a symmetrical perspective but also a practical one;
when using the `decode` API you're more likely to want to know the next
offset to slice or continue decoding from, than you are to need the
bytes decoded. However this is a...

BREAKING CHANGE:

 - `decode()` and `decode32()` now return a tuple of the encoded number,
   and the _new offset_ as opposed to the _bytes decoded_. To get the
   bytes decoded, simply negate the original given `offset` from the
   returned number.